requirements.txt
Flask==2.3.3
google-cloud-aiplatform==1.60.0

google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.2.0

google-generativeai==0.1.0  # Replace with the actual version you are using
